Project5/requirements.txt
openai
langchain
streamlit
streamlit-chat
pinecone-client
python-dotenv
tiktoken
